+++ app/src/main/java/com/konkuk/medicarecall/data/repository/ElderIdRepository.kt
@@ -1,8 +1,8 @@
 package com.konkuk.medicarecall.data.repository
 
 interface ElderIdRepository {
-    suspend fun updateElderIds(elderIdMap: Map<Int, String>)
-    suspend fun updateElderId(elderId: Int, name: String)
+    suspend fun updateElderIds(elderIdMap: Map<Long, String>)
+    suspend fun updateElderId(elderId: Long, name: String)
     suspend fun clearElderIds()
-    suspend fun getElderIds(): Map<Int, String>
+    suspend fun getElderIds(): Map<Long, String>
 }
