+++ app/src/main/java/com/konkuk/medicarecall/data/repository/MentalRepository.kt
@@ -4,5 +4,5 @@ import com.konkuk.medicarecall.domain.model.Mental
 import java.time.LocalDate
 
 interface MentalRepository {
-    suspend fun getMental(elderId: Int, date: LocalDate): Result<Mental>
+    suspend fun getMental(elderId: Long, date: LocalDate): Result<Mental>
 }
