+++ app/src/main/java/com/konkuk/medicarecall/data/repository/StatisticsRepository.kt
@@ -3,5 +3,5 @@ package com.konkuk.medicarecall.data.repository
 import com.konkuk.medicarecall.domain.model.StatisticsData
 
 interface StatisticsRepository {
-    suspend fun getStatistics(elderId: Int, startDate: String): Result<StatisticsData>
+    suspend fun getStatistics(elderId: Long, startDate: String): Result<StatisticsData>
 }
