+++ app/src/main/java/com/konkuk/medicarecall/ui/model/ElderSubscription.kt
@@ -1,7 +1,7 @@
 package com.konkuk.medicarecall.ui.model
 
 data class ElderSubscription(
-    val elderId: Int,
+    val elderId: Long,
     val name: String,
     val plan: String,
     val price: Int,
